requirements.txt
Flask
Flask-Session
Flask-SocketIO
psycopg2
SQLAlchemy
requests
python-dotenv
eventlet
